Create shared examples for follows
