+++ decidim-core/lib/decidim/core/test.rb
@@ -18,3 +18,4 @@ require "decidim/core/test/shared_examples/paginated_resource_examples"
 require "decidim/core/test/shared_examples/errors"
 require "decidim/core/test/shared_examples/scope_helper_examples"
 require "decidim/core/test/shared_examples/user_localised_email_examples"
+require "decidim/core/test/shared_examples/follows_examples"
